# api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PointMarket Backend API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/api/items": {
            "get": {
                "tags": [
                    "Item"
                ],
                "description": "現在販売中（= status が on_sale）の商品の一覧を取得する",
                "operationId": "getItems",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ItemResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items/{item_id}": {
            "get": {
                "tags": [
                    "Item"
                ],
                "description": "現在販売中（= status が on_sale）の商品で指定されたIDの商品を取得する",
                "operationId": "getItemById",
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items/{item_id}/buy": {
            "post": {
                "tags": [
                    "Item"
                ],
                "description": "現在販売中（= status が on_sale）の商品で指定されたIDの商品を、自身の保有するポイントで購入する",
                "operationId": "buyItemById",
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/my/items": {
            "get": {
                "tags": [
                    "MyItem"
                ],
                "description": "自分の登録している商品の一覧を取得する",
                "operationId": "getMyItems",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MyItemResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "MyItem"
                ],
                "operationId": "postMyItem",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MyItemSaveRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MyItemResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        },
        "/api/my/items/{item_id}": {
            "delete": {
                "tags": [
                    "MyItem"
                ],
                "description": "自分が登録した商品の削除",
                "operationId": "deleteMyItem",
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            },
            "patch": {
                "tags": [
                    "MyItem"
                ],
                "description": "自分が登録した商品の更新",
                "operationId": "patchMyItem",
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MyItemSaveRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MyItemResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "postUserLogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        },
        "/api/users/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "getMe",
                "responses": {
                    "200": {
                        "description": "AOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        },
        "/api/users/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "postUserRegister",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegisterPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "MyItemSaveRequest": {
                "required": [
                    "name",
                    "selling_price_point",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "not_on_sale",
                            "on_sale"
                        ]
                    },
                    "selling_price_point": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "UserLoginPostRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "test-user@point-market.jp"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "UserRegisterPostRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ItemResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "description": "品名",
                        "type": "string"
                    },
                    "description": {
                        "description": "商品詳細",
                        "type": "string"
                    },
                    "status": {
                        "description": "状態",
                        "type": "string"
                    },
                    "selling_price_point": {
                        "description": "販売価格（ポイント）",
                        "type": "integer"
                    },
                    "seller_user_id": {
                        "description": "販売しているユーザーのID",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "登録日時",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "最終更新日時",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "MyItemResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "description": "品名",
                        "type": "string"
                    },
                    "description": {
                        "description": "商品詳細",
                        "type": "string"
                    },
                    "status": {
                        "description": "状態",
                        "type": "string"
                    },
                    "selling_price_point": {
                        "description": "販売価格（ポイント）",
                        "type": "integer"
                    },
                    "buyer_user_id": {
                        "description": "購入したユーザーのID",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "登録日時",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "最終更新日時",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "description": "品名",
                        "type": "string"
                    },
                    "points": {
                        "description": "保有ポイント",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-XSRF-TOKEN": {
                "type": "apiKey",
                "name": "XSRF-TOKEN",
                "in": "cookie"
            }
        }
    }
}